MAC005-19 | edit supplier -> create contact(datatable, create, delete) - MAC005-21 | funcionalidad de borrar para módulo supplier.

+++ database/migrations/2017_08_22_154357_create_additional_charges_table.php
@@ -18,7 +18,7 @@ class CreateAdditionalChargesTable extends Migration
             $table->string('concept');
             $table->string('collect_prepaid');
             $table->string('import_export');
-            $table->integer('amount');
+            $table->double('amount', 15, 2);
             $table->string('currency');
             $table->date('last_updated');
             $table->enum('charge_type', ['BL', 'Container', 'Others']);
